Adding an Item successfully posts to backend DB & updates map instantly
Switched to position field to 'geo : [lat, lng]' to be compliant with the Database Schema. This format is for mongoose geospatial query support. React-google-maps requires that its map markers have the object format, but its easier to just change the properties on the item while adding them to the markers array