Moved AddItem to its own component, placed udner map w/ flexbox
This change should provide a better user experience, because the old locaiton of the AddItem form was below the feed of nearby items, and it would move around and sometimes go below the fold. Used flexbox column for the elements in the sidebar (map and AddItem form) and justify-content flex-end so that the form's container will always be big enough, and the map will stretch to take up the rest of the space alloted. Also tidyed up the Map presentational component