Implemented Item deletion for owner of the item listing
If you are the person who listed an item, you will see a red 'X' button on the item card. Clicking this X dispatches an async action that removes the item from the database by its ID, then resolves with the deleted item if successful. The item reducer then filters out the deleted item from next iteration of state and the UI updates immediately. Considering to add an 'are you sure?' type modal before dispatching the delete