Updated React App for Surge deployment
Surge is static hosting. With Create React App, we can set different environment variables for dev vs prod, by making a .env.development and .env.production file. Here, I am placing an env var in both files, dev has the local URL and prod the production URL on Vultr (for now). REACT_APP_SERVER_HOST_URL. The REACT_APP_ preface makes it work with create-react-app so that after building, the production env var will be used. I set the 'host' variable in the APIManager.js in utils/ as the REACT_APP_SERVER_HOST_URL, and this is the address that API requests will be made to